add custom textarea validation && fix bugs

+++ client/src/providers/dataProvider.ts
@@ -6,6 +6,12 @@ import { $api } from "../api/http";
 
 import qs from "qs";
 
+const INVALID_RESOURCE_OPERATIONS = {
+	create: ["categories"],
+	delete: ["categories"],
+	deleteMany: ["categories"]
+};
+
 // @ts-ignore. don't need to use updateMany / getManyReference
 const provider: DataProvider = {
 	getOne: async (res, par) => {
@@ -44,17 +50,19 @@ const provider: DataProvider = {
 			data: { data, total }
 		} = await $api.get(url);
 
-		console.log();
-
 		const formatted = formatObjId(data);
 
 		return { data: [...(Array.isArray(formatted) ? formatted : [])], total };
 	},
 	create: async (res, par) => {
-		const {
-			data: { data }
-		} = await $api.post(`/${res}`, { ...par.data });
-		return { data: formatObjId(data) };
+		if (!INVALID_RESOURCE_OPERATIONS.create.includes(res)) {
+			const {
+				data: { data }
+			} = await $api.post(`/${res}`, { ...par.data });
+			return { data: formatObjId(data) };
+		}
+
+		return Promise.reject("You can't create this resource");
 	},
 	update: async (res, par) => {
 		const {
@@ -63,19 +71,27 @@ const provider: DataProvider = {
 		return { data: formatObjId(data) };
 	},
 	delete: async (res, par) => {
-		const {
-			data: { data }
-		} = await $api.delete(`/${res}/${par.id}`);
-		return { data: formatObjId(data) };
+		if (!INVALID_RESOURCE_OPERATIONS.delete.includes(res)) {
+			const {
+				data: { data }
+			} = await $api.delete(`/${res}/${par.id}`);
+			return { data: formatObjId(data) };
+		}
+
+		return Promise.reject("You can't delete this resource");
 	},
 	deleteMany: async (res, par) => {
-		const url = `/${res}/many/${JSON.stringify(par.ids)}`;
+		if (!INVALID_RESOURCE_OPERATIONS.deleteMany.includes(res)) {
+			const url = `/${res}/many/${JSON.stringify(par.ids)}`;
 
-		const {
-			data: { ids }
-		} = await $api.delete(url);
+			const {
+				data: { ids }
+			} = await $api.delete(url);
 
-		return { data: formatObjId(ids) };
+			return { data: formatObjId(ids) };
+		}
+
+		return Promise.reject("You can't delete this resources");
 	}
 };
